Bootrom Tools: Revising sign-tftf to use standard test and real signature keys
Now that we have official keys for signing, they need to be integrated into
both the bootrom tree and the build process. This revises the signing tools
to use a pair of standard (test and real) key files. The original --key
parameter is retained for extraordinary cases.

Testing Done:
Bug Number:
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>