Disable Gear Changing test code
We implemented the Gear Changing code for testing, but did not want
to include it in the drop to Toshiba since there was not enough time
for back and forth simulation. But for some reason it got slipped in.
And the simulation from Toshiba indicates it is causing some problem.
While it is nice to figure out what the problem is, we need to focus
on normal scenarios first. So fix the build script to make sure it
does not get in by default.

Testing Done:   The gear changing code is not compiled in/
Bug Number:     Trivial
Review URL:     https://reviews.leaflabs.com/r/1863/
Reviewed by:    Eli Sennesh <esennesh@leaflabs.com>
                Morgan Girling <girling_morgan@projectara.com>

Signed-off-by:  Jun Li <li_jun@projectara.com>