Bootrom Tools: Test Suite: Modified for response files and FTDI integration
The FTDI integration, as well as our current test paradigm, uses response
files (.rsp) instead of pass/fail strings. This commit converts the test
execution (run-bootrom-tests) and the test-suite creation
(create-bootrom-test-suite) to this new form.

The above neccessitates also changing the test suite script (*.tss),
eliminating the pass/fail strings, adding the response files, and
disambiguating which ffff to which the patch instructions apply.

NOTE: This is a mostly-standalone set of changes, but it does require
the C-base FTDI (in another review) to run.

Testing Done: Lots of test runs standalone and with the FTDI library.
Bug Number: SW-1039
Review URL: https://reviews.leaflabs.com/r/1891/
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>