Bootrom Tools: Automated Testing: FTDI support for test automation
This is a port of Jun's "semi-automated test" code, plundered for
its FTDI support, plus Jun's earlier "~/spi" folder, which created a stand-
alone HAPS daughterboard SPIROM flashing program. This is a comparatively
rough port of the above 2 standalone folders, adapting it from build scripts
to a simple Makefile, making it interchangeable between Jun's environment
and mine, and changing the arguments to "haps_test" to integrate it with
the changes to the Python wrapper (see: review #1891). Rather than delaying
deployment to refine the code, this is in a "known-to-work" state that others
can use while cleanup and refinements continue.

On particularly unfortunate aspect of this check-in is the inclusion of the
entire LibMPSSE-SPI_source folder. Since the semi-automated test shells out
to run the Flash program, it was best to bring in into this block of work.
Unfortunately, it uses a "single" file from within that folder
(LibMPSSE-SPI_source/LibMPSSE-SPI/LibMPSSE/Build/Linux/libMPSSE.a). It will
take some time to pare away the unneeded parts of that folder or integrate the
SPIROM functionality directly into the rest of the FTDI code. For now, please
ignore that folder (which I believe to be untouched) for now.

Testing Done: verified that it can run a complete test (flash one or both
daughterboards, download and boot one or both daughterboards, and capture the
debug output from the "bridge" daughterboard. This has been repeated with
various configurations, images, etc.
Bug Number: SW-1039
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>