Bootrom Tools: Bootrom Test: Eliminated env. vars from run-bootrom-tests
Run-bootrom-tests used environment vars to read/write persistent state
(the last server and DUT flash files reflashed) and the path to the
"haps_test" helper app. Unfortunately, it also needs to be run as root,
at which point, it inherits root's environment, rendering the environment
variable usage moot.

This commit uses hidden files for the 2 persistent state variables, and a
required command-line arg for the ftdi path.

Testing Done: verified that the file-based persistent variables worked as
expected, and that the ftdi-path is properly imported via the command line
and distributed to the code that requires it.
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>