Bootrom Tools: Improved arg parser in C-based tools
Changed the mechanism for specifying multiple long-names from a comma-
separated string that gets edited, to a pointer to an array of strings.
This allows the set of long names to be listed in the usage help strings.
(Incorporates Jun's review feedback)

Testing Done: ran the various tools with good and bad parameters
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2077/
Reviewed by: Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>