Bootrom Tools: Fix number of reserved cells in ffff (Python)
Due to a misinterpretation, the number of Reserved entries in an FFFF was
made variable instead of fixed. This commit makes the number fixed and brings
it in-line with the "C" version.

Testing Done: verified the layout of the generated files via the "C" version
of the tools
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>