Bootrom Tools: Added auto-name to create-tftf
I hadn't noticed that someone had added an auto-generated default output file
name to the Python version. I have ported this, and as much of "validate_args"
as makes sense from the python to the C code. In testing, I discovered that
--ara_stage (left over from early bootrom development) conflicted with --type
and deleted the --ara_stage support (which only affected the auto-generated
filename).

Testing Done: verified that it handles the parameters correctly.
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>