Bootrom Tools: Add "ignore address" to create-tftf (C)
Integrated the "ignore address" for signature and certificate sections into
create-tftf. This commit ensures that both C and Python versions handle it.

Testing Done: Tested with a --signature file --load address, confirmed that
it flagged the warning. Tested with a --code following a --signature,
confirmed that it was flagged as an error.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2151/
Reviewed by: Jun Li

Signed-off-by:  Morgan Girling <morgang@bsquare.com>