Bootrom Tools: Misc. Python tweaks
Trimmed the firmware package name to fit in 48 bytes c/w the trailing null
(previously used 49 bytes), and issue a warning message to the user when it
truncated the name. The --map option prints the correct offsets for the
section data.
Similar fix to ffff.py

Testing Done: co-verified with the C version
Bug Number: trivial
Review URL:
Reviewed by:

Signed-off-by:  Morgan Girling <morgang@bsquare.com>