Bootrom Tools: Auto-balance short MID/VID/PID (Python)
Vendors have been told to use 15-bit MIDs, VIDs and PIDs, but the bootrom
requires Hamming-balanced values. The workaround is to copy the 1s
compliment of the lower 16 bits into the upper 16 bits. This commit
auto-balances values <= 0x7fff (and issues a warning of the upconversion)
and checks all other values for a Hamming weight of 16. (As always, zero
is a special case which is simply passed through.)

Testing Done: ran it against values of 0x1234, 0x5678, 0x9abc, 0xdef0.
The first 2 were upconverted, and the 3rd was flagged as an error. Since
the first error stops processing, the final one was not checked.
Bug Number: trivial
Review URL: https://reviews.leaflabs.com/r/2220/
Reviewed by: Eli Sennesh

Signed-off-by:  Morgan Girling <morgang@bsquare.com>